prologue: add PHY managements on tests start
Allow PHY parameters to be changed before starting tests. This is useful
for testing NICs not only with default PHY parameters, but also with those
set manually, which expands the testing possibilities. Previously, it was
not possible to test at the minimum allowable speed or in a half-duplex state.
This patch fixes this.

OL-Redmine-Id: 14619
Signed-off-by: Daniil Byshenko <daniil.byshenko@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>